Extract chat types from scenario types.
