+++ src/hooks/usePersistedState.ts
@@ -1,5 +1,5 @@
 import createPersistedState from 'use-persisted-state';
-import { ChatDialog } from "../types/ScenarioTypes";
+import { ChatDialog } from "../types/ChatTypes";
 
 const useChatState = createPersistedState<ChatDialog[]>("chatState");
