+++ src/scenarios/scenario1/scenario1.json
@@ -1,10 +1,22 @@
 [
   {
-    "chats": [{ "sender": "George Clooney", "message": "Hello, World!" }],
+    "chats": [
+      {
+        "senderId": "George Clooney",
+        "message": "Hello, World!",
+        "timestamp": "2022-07-05T00:00:00.000Z"
+      }
+    ],
     "endRepoID": "end1"
   },
   {
-    "chats": [{ "sender": "George Clooney", "message": "Good Job!" }],
+    "chats": [
+      {
+        "senderId": "George Clooney",
+        "message": "Good Job!",
+        "timestamp": "2018-01-01T00:00:12.531Z"
+      }
+    ],
     "endRepoID": "end2"
   }
 ]
