+++ src/types/ChatTypes.ts
@@ -1,0 +1,5 @@
+export type ChatDialog = {
+  senderId: string;
+  message: string;
+  timestamp: Date;
+};
