+++ src/types/ScenarioTypes.ts
@@ -1,9 +1,6 @@
+import { ChatDialog } from "./ChatTypes";
+
 export type ScenarioSegment = {
   chats: ChatDialog[];
   endRepoID: string /* ID of the comparing repository  */;
-};
-
-export type ChatDialog = {
-  sender: string;
-  message: string;
 };
